update: refactor stable_diffusion module
