+++ wandb_addons/diffusers/callbacks/stable_diffusion/__init__.py
@@ -1,0 +1,5 @@
+from .text_to_image import StableDiffusionCallback
+from .image_to_image import StableDiffusionImg2ImgCallback
+
+
+__all__ = ["StableDiffusionCallback", "StableDiffusionImg2ImgCallback"]
